pyproject.toml
[tool.poetry]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = ["Carlos Cezar Negreiros <cezarnegreiros07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = ">=5.1.6,<6.0.0"
psycopg2-binary = "^2.9.10"
python-decouple = "^3.8"
djangorestframework = "^3.16.1"
requests = "^2.32.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
